% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_methods.R
\name{print.empbary_result}
\alias{print.empbary_result}
\title{Print an empirical barycenter result}
\usage{
\method{print}{empbary_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"empbary_result"}.}

\item{...}{Unused.}
}
\description{
Print an empirical barycenter result
}
