% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_methods.R
\name{print.mcbary_result}
\alias{print.mcbary_result}
\title{Print an MCB barycenter result}
\usage{
\method{print}{mcbary_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mcbary_result"}.}

\item{...}{Unused.}
}
\description{
Print an MCB barycenter result
}
